src/test/java/io/github/svaponi/HelloControllerTest.java
package io.github.svaponi;

import io.github.svaponi.resource.AnyResource;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mockito;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.HttpEntity;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.junit4.SpringRunner;

import javax.servlet.http.HttpServletRequest;

import static org.hamcrest.Matchers.instanceOf;
import static org.junit.Assert.assertThat;

@RunWith(SpringRunner.class)
@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
@ActiveProfiles("unit-test")
public class HelloControllerTest {

    private final HelloController controller;

    public HelloControllerTest() {
        this.controller = new HelloController();
    }

    @Test
    public void root() throws Exception {
        final HttpEntity<?> response = controller.root(Mockito.mock(HttpServletRequest.class));
        assertThat(response.getBody(), instanceOf(AnyResource.class));
    }
}